pyproject.toml
[tool.poetry]
name = "sdk-load-test"
version = "0.1.0"
description = "Package for load testing the langsmith sdk"
authors = ["Robert Xu <xuro@langchain.dev>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
langsmith = "^0.4.1"
langchain = "^1.0.0"
langchain_openai = "^1.0.0"
orjson = "^3.10.14"
python-dotenv = "^1.0.1"
pathlib = "^1.0.1"
humanize = "^4.11.0"
pandas = "^2.3.0"
packaging = "^25.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
